% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sharedClonotypes.R
\docType{data}
\name{shared_clonotypes}
\alias{shared_clonotypes}
\title{Shared clonotypes}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 413 rows and 5 columns.
}
\usage{
data(shared_clonotypes)
}
\description{
\code{data.table} of the clonotypes that are found in both
conditions MUT and WT in the 0606T1 data set.
}
\examples{
\dontrun{

## count TRA/TRB
clono_freq <- colData(sce.filt)[, c("Sample" ,"cdr3s_aa")] \%>\%
   as.data.frame \%>\% data.table(., keep.rownames = TRUE) \%>\%
    .[!is.na(cdr3s_aa), .N, c("cdr3s_aa","Sample")]
setorder(clono_freq, N)

## formatting the TRA/TRB notations
## will only work if there's just one TRA
ct <- dcast(clono_freq, cdr3s_aa ~ Sample, value.var = "N") \%>\%
 .[!is.na(MUT.0606T1) & !is.na(WT.0606T1)]

ct[, TRA := gsub(";*TRB:[A-Z]+", "", cdr3s_aa)]
ct[, TRA := ifelse(TRA == "", NA, TRA)]
ct[, TRB := gsub(".*(TRB:[A-Z]+)", "\\\\1", cdr3s_aa)]
ct[, TRB := ifelse(grepl("^TRA", TRB), NA, TRB)] # if only TRB was present,
 I need to fill in the NA
 setorder(ct, -MUT.0606T1, -WT.0606T1 )
 shared_clonotypes <- copy(ct)
}

}
\seealso{
\code{clonotype_ids}
}
\keyword{datasets}
